Add functions for increment and decrement confirmed requests

+++ ewm-main-service/src/main/java/ru/practicum/event/dto/NewEventDto.java
@@ -35,7 +35,7 @@ public class NewEventDto {
     @NotNull
     private LocationDto location;
     private Boolean paid = false;
-    private Integer participantLimit = 0;
+    private Long participantLimit = 0L;
     private Boolean requestModeration = true;
 
     @NotEmpty
